requirements/local.txt
-r base.txt

# Testing
# ------------------------------------------------------------------------------
mypy==1.6.0  # https://github.com/python/mypy
django-stubs==4.2.3  # https://github.com/typeddjango/django-stubs
pytest==7.4.2  # https://github.com/pytest-dev/pytest

# Code quality
# ------------------------------------------------------------------------------
flake8==6.1.0  # https://github.com/PyCQA/flake8
flake8-isort==6.1.1  # https://github.com/gforcada/flake8-isort
coverage==7.3.2  # https://github.com/nedbat/coveragepy
black==23.7.0  # https://github.com/psf/black
djlint==1.32.1  # https://github.com/Riverside-Healthcare/djLint
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pre-commit==3.5.0  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
django-debug-toolbar==4.2.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
factory-boy==3.3.0  # https://github.com/FactoryBoy/factory_boy
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django
djangorestframework==3.14.0
django-filter==23.3
django-tables2
django-filter
